requirements.txt
python-dotenv
openai
pillow
flask
flask-cors